# docs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Server.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/job-application": {
      "post": {
        "tags": [
          "JobApplication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobApplicationCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobApplicationCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobApplicationCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "JobApplication"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/job-application/bulk": {
      "post": {
        "tags": [
          "JobApplication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobApplicationsCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobApplicationsCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobApplicationsCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/job-application/{id}/feedback": {
      "post": {
        "tags": [
          "JobApplication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobApplicationFeedbackCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobApplicationFeedbackCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobApplicationFeedbackCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/job-application/{applicationId}/feedback/{feedbackId}": {
      "put": {
        "tags": [
          "JobApplication"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "feedbackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditJobApplicationFeedbackCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EditJobApplicationFeedbackCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EditJobApplicationFeedbackCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "JobApplication"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "feedbackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/job-application/{id}/move-status": {
      "put": {
        "tags": [
          "JobApplication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveJobApplicationStatusCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveJobApplicationStatusCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MoveJobApplicationStatusCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/job-application/{id}": {
      "delete": {
        "tags": [
          "JobApplication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "JobApplication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/job-application/for-job-opening/{jobOpeningId}": {
      "get": {
        "tags": [
          "JobApplication"
        ],
        "parameters": [
          {
            "name": "jobOpeningId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateJobApplicationCommand": {
        "type": "object",
        "properties": {
          "candidateId": {
            "type": "string",
            "format": "uuid"
          },
          "jobOpeningId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "CreateJobApplicationFeedbackCommand": {
        "type": "object",
        "properties": {
          "jobApplicationId": {
            "type": "string",
            "format": "uuid"
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "skillFeedbacks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillFeedbackDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateJobApplicationsCommand": {
        "type": "object",
        "properties": {
          "applications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobApplicationDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EditJobApplicationFeedbackCommand": {
        "type": "object",
        "properties": {
          "jobApplicationId": {
            "type": "string",
            "format": "uuid"
          },
          "feedbackId": {
            "type": "string",
            "format": "uuid"
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "skillFeedbacks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillFeedbackDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "JobApplicationDTO": {
        "type": "object",
        "properties": {
          "candidateId": {
            "type": "string",
            "format": "uuid"
          },
          "jobOpeningId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "JobApplicationStatus": {
        "enum": [
          "Applied",
          "Shortlisted",
          "Interviewed",
          "Offered",
          "Hired",
          "OnHold",
          "Rejected"
        ],
        "type": "string"
      },
      "JobApplicationStatusActions": {
        "enum": [
          "UnHold",
          "RollBack"
        ],
        "type": "string"
      },
...
      "MoveJobApplicationStatusCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "moveTo": {
            "$ref": "#/components/schemas/JobApplicationStatus"
          },
          "action": {
            "$ref": "#/components/schemas/JobApplicationStatusActions"
          }
        },
        "additionalProperties": false
      },
...
      "SkillFeedbackDTO": {
        "type": "object",
        "properties": {
          "skillId": {
            "type": "string",
            "format": "uuid"
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "assessedExpYears": {
            "type": "number",
            "format": "float",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}